pm: improvements to install mostly
None of them use UTF runes that are
not likely to exist in some fonts.
Install accepts -v to report the files installed
while it goes.
Install pulls if that's not done, instead of
asking the user to do so.
Install relies on tar T option instead of
going file by file adjusting modes/gids.

R=nix-dev, john
CC=nix-dev
http://codereview.appspot.com/5000045