pm: multiple servers and sources in /sys/src This is a first attempt at having packages keep their source at /sys/src. To symplify, I left a root.tgz for devel/indent at lsub.org, so that if you list that server before, you'd get a package with the tar using /sys/src/cmd/... Now the changes...
lib/ndb/local has entries that describe a list of
servers for packages (tried in order).
pm/pull has changed to try that list. Still defaults
to the default, and honors the variable (which can now
be alist) if set.
pm/list has a new flag -l that lists the files for
packages installed. To be used like pm/list -il devel/indent
(or just pmlist -il).
pm/install changed because it now says nothing regarding
the source of the packages (either the caller or pull
fixes that).
the mkfile for sys/src/cmd includes in the list of std. dirs
only those that have a file named "mkfile" in them.
I think it's reasonable to assume that such packages obey
the std. rules (i.e., have the std targets).
If that's not the case, we could simply remove any mkfile
to mkfile.checkme or whatever.

I think that was all. It would be great if you could try
before any submit. As I said, edit the ndb entry so that
the first server listed is just "lsub.org" (but you can
only retrieve devel/indent from there).

R=john, nemo
CC=nix-dev
http://codereview.appspot.com/5040043