fixes: more ignores, codereview.plan9, acpi
This adds more directories to .hgignore, like
srv, fd, mnt, ... in case anyone uses hg within
Plan 9. Doing this, and mounting the root
again using a new connection, and using that
dir to run hg seems to be enough to get hg happy.
Otherwise it would enter directories that have devices...

the coderevivew.py for plan 9 has been changed to
ignore /sys/log in changes, and also not to ask the
user when it sees a new file. Otherwise, you'd get a
prompt for each file you have create but do not want
to submit.

physalloc.c has a fix so that if acpi reports as a
domain id a number < 0 the kernel would panic
instead of crashing in some other way. It should
help detect problems.

R=nix-dev, rminnich
CC=nix-dev
http://codereview.appspot.com/5082042