segment: size miscalcuation
the size in sg.size is a number of pages, not
a size in bytes.  using a size in bytes caused sg.size
to saturate to zero, thus causing telnet(1) to fail.

this isn't a complete fix.  i think sg.size should be
in bytes and thus a uintptr.  then we don't need
to tie page sizes to segment sizes so strongly.

R=nixiedev
CC=nix-dev
http://codereview.appspot.com/5435046

Committer: Noah Evans <noah.evans@gmail.com>