telnet.c: don't ignore segattach failure
1.  don't ignore segattach failure.  in addition, it's not
necessary to tell segattach the virtual address to attach
the segment at.   (and run the risk of picking a bad
address.)  as per segattach(2) when given a virtual address
of 0, segattach will find a hole in the addres space and
attach it there.

2.  if posting to /srv to allow later connections, daemonize
(fork) so the caller doesn't need to remember to background a
task that always runs in the background.

3.  supress fiddling with /dev/consctl if posting the
connection to /srv.  that's the wrong consctl.  the console
doing the connecting is attaching to the file in /srv,
and we don't have access to his consctl.  he'll need to
make his own arrangements.

rawon() and rawoff() were merged to consctl() because the
variable bit (the message "rawon" or "rawoff") was much
smaller than the common bit (all the checks and opening
the console).

R=nixiedev, noah.evans
CC=nix-dev
http://codereview.appspot.com/5432046

Committer: Noah Evans <noah.evans@gmail.com>