devacpi: bogus xdsdt in ACPI table
A bogus xdsdt address in the ACPI table seemed
to be rendering my system unbootable. Using the
dsdt seems to have resolved the issue. In the
general case, I have followed the ACPICA: match
the 32-bit address with the 64-bit address, and
prefer the 32-bit address. Unlike facs vs. xfacs
(where one is necessarily 0), the ACPI spec sheet
(4.0a) does not specify what happens between
dsdt vs xdsdt. This case seems to show that in
implementation, it is not necessary that one is 0.

R=nix-dev, nemo, john
CC=nix-dev
http://codereview.appspot.com/5556048

Committer: John Floren <john@jfloren.net>