plumber: fix memory leaks and arm arch crashes
there are a few sloppy bits in the plumber.  upas tends
to find these, and plumber didn't have enough stack for
arm.

the biggest change was to get rid of the ill-advised static
buffer in libplumb, which is not thread safe.  adding a
mutex caused some deadlocks, iirc with upas/fs.

this was the message attached to the patch in plan 9
(maybe/maybe/plumbmem)

fix crash due to timing bug in libplumb as here
http://9fans.net/archive/2009/04/103
http://9fans.net/archive/2009/04/121
http://9fans.net/archive/2009/04/124
http://9fans.net/archive/2009/04/126
also fix a few trivial memory leaks in plumber.

R=nixiedev, nemo
CC=nix-dev
http://codereview.appspot.com/5756060

Committer: Francisco J Ballesteros <nemo.mbox@gmail.com>