kexec: Kexec ported to the latest version of Nix.  Untested, but this should be the minimum set of changes to get everything to work.
Kexec provides for the in kernel execution (of relinked)
standard plan9 executables.  It's not yet what you'd expect
though.  traditional libc will not work, the current kludge is
to replace the system calls in libc with function pointers to
the relevant kernel functions (use nm on your compiled kernel)
or to provide wrappers which marshal the system calls directly
(still using nm to find the syscalls).  Libdynld and a
modified devdynld are probably a better interface than the
current approach.

R=rminnich, nemo.mbox, nixiedev, quanstro
CC=nix-dev
http://codereview.appspot.com/5936051