Remove Ghostscript related LICENSEs
In the stripped down version of Nix, there are no Ghostscript pieces
left, neither the program (afpl) nor the fonts (gpl). These can go.
(And for OS distributions, other unencumbered fonts are available, the
TeX related ones and the Hershey ones for example.)

For the LICENSE and various NOTICE found in subdirs:

The original code has been re-licensed to MIT with this:

Copyright 2021 Plan 9 Foundation

Permission is hereby granted, free of charge, to any person obtaining
a copy of this software and associated documentation files (the
"Software"), to deal in the Software without restriction, including
without limitation the rights to use, copy, modify, merge, publish,
distribute, sublicense, and/or sell copies of the Software, and to
permit persons to whom the Software is furnished to do so, subject to
the following conditions:

The above copyright notice and this permission notice shall be
included in all copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY
CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.

But this concerns the Plan9 Lucent Public License code (with the
problem of the date in the above).

For Nix and the pieces with NOTICE, this has to be viewed at the
very least by the Nix original authors and it will further remain to
indicate what piece is under what license.

So LICENSE and the various NOTICE are left untouched in the present
commit.

Signed-off-by: Thierry LARONDE <tlaronde@kergis.com>