Cleaning: removing useless files and including words in NOTES
NOTICE, sys/src/NOTICE: rm. These two remaining NOTICE, indicating
a copyright of 2002 (so before Nix was launched) with the Lucent
License and hence relative to the plan9 distribution. But this has
been relicensed. So obsolete (and even in contradiction with
LICENSE).

sys/src/nix/k10/Linux: rm. Indicating that the Linux support has been
removed and pointing to a dumb that is nowhere present.

sys/src/nix/words (rm), NOTES (modified): contained some explanations
about the team, some features and how to start -> incorporated
(slightly edited) in NOTES at the root of the repository.

Signed-off-by: Thierry LARONDE <tlaronde@kergis.com>